requirements.txt
Flask==3.0.3
moviepy
pillow
qrcode
apscheduler
gunicorn
imageio-ffmpeg
cloudinary
requests